// React/app/GerAdv_TS/NEPalavrasChaves/Interfaces/interface.NEPalavrasChaves.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface INEPalavrasChaves {
// 202501251
    id: number;
 
	nome: string,
}

export interface INEPalavrasChavesFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface INEPalavrasChavesIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}